5062 - organize code into sub-directories
