+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/archive/oldcode/Red20SecAuto.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.ftc_2021_2022.oldcode;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022.archive.oldcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
