+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/archive/oldcode/RedWarehouseForward.java
@@ -27,15 +27,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.ftc_2021_2022.oldcode;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022.archive.oldcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
-
-import org.firstinspires.ftc.robotcontroller.external.samples.HardwarePushbot;
 
 /**
  * This file illustrates the concept of driving a path based on time.
@@ -58,9 +56,9 @@ import org.firstinspires.ftc.robotcontroller.external.samples.HardwarePushbot;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@Autonomous(name="RedWarehouse", group="Autonomous")
+@Autonomous(name="RedWarehouseFoward", group="Autonomous")
 @Disabled
-public class RedWarehouse extends LinearOpMode {
+public class RedWarehouseForward extends LinearOpMode {
 
     /* Declare OpMode members. */
     private ElapsedTime runtime = new ElapsedTime();
@@ -111,21 +109,36 @@ public class RedWarehouse extends LinearOpMode {
 
         runtime.reset();
         while (opModeIsActive() && (runtime.seconds() < 2.5)) {
-//            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-//            telemetry.update();
+            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
+            telemetry.update();
         }
 
         // Step 2:  Stop
         strafeRight(0); //Stops motors
-        arm.setPower(0);
+        //forward (0.1);
+        leftTop.setPower(0.2);
+        leftBottom.setPower(0.2);
+        rightTop.setPower(0.2);
+        rightBottom.setPower(0.2);
 
-//        telemetry.addData("Path", "Complete");
-//        telemetry.update();
+        runtime.reset();
+        while (opModeIsActive() && (runtime.seconds() < 1.0)) {
+            telemetry.addData("Path", "Leg 2: %3.5f S Elapsed", runtime.seconds());
+            telemetry.update();
+        }
+
+        //forward(0);
+        leftTop.setPower(0);
+        leftBottom.setPower(0);
+        rightTop.setPower(0);
+        rightBottom.setPower(0);
+
+        telemetry.addData("Path", "Complete");
+        telemetry.update();
         sleep(1000);
     }
 
     public void strafeRight(double speed) {
-
         double rb_speed_var = 0;
         if (speed > 0) {
             rb_speed_var = 1 / speed;
@@ -138,6 +151,13 @@ public class RedWarehouse extends LinearOpMode {
         leftBottom.setPower(-speed);
         rightTop.setPower(-speed);
         rightBottom.setPower(rb_speed);
+    }
+
+    public void forward(double speed) {
+        leftTop.setPower(speed);
+        leftBottom.setPower(speed);
+        rightTop.setPower(speed);
+        rightBottom.setPower(speed);
     }
