+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/archive/trish_sandbox/RobotConfigTrish.java
@@ -27,144 +27,136 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.ftc_2021_2022;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022.archive.trish_sandbox;
 
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
+import org.firstinspires.ftc.robotcore.external.Telemetry;
+
 /**
- * This file illustrates the concept of driving a path based on time.
- * It uses the common Pushbot hardware class to define the drive on the robot.
- * The code is structured as a LinearOpMode
+ * This is NOT an opmode.
  *
- * The code assumes that you do NOT have encoders on the wheels,
- *   otherwise you would use: PushbotAutoDriveByEncoder;
+ * This class can be used to define all the specific hardware for a single robot.
+ * In this case that robot is a Pushbot.
+ * See PushbotTeleopTank_Iterative and others classes starting with "Pushbot" for usage examples.
  *
- *   The desired path in this example is:
- *   - Drive forward for 3 seconds
- *   - Spin right for 1.3 seconds
- *   - Drive Backwards for 1 Second
- *   - Stop and close the claw.
+ * This hardware class assumes the following device names have been configured on the robot:
+ * Note:  All names are lower case and some have single spaces between words.
  *
- *  The code is written in a simple form with no optimizations.
- *  However, there are several ways that this type of sequence could be streamlined,
- *
- * Use Android Studios to Copy this Class, and Paste it into your team's code folder with a new name.
- * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
+ * Motor channel:  Left  drive motor:        "left_drive"
+ * Motor channel:  Right drive motor:        "right_drive"
+ * Motor channel:  Manipulator drive motor:  "left_arm"
+ * Servo channel:  Servo to open left claw:  "left_hand"
+ * Servo channel:  Servo to open right claw: "right_hand"
  */
+public class RobotConfigTrish
+{
+    /* Public OpMode members. */
+    public DcMotor leftTop = null;
+    public DcMotor leftBottom = null;
+    public DcMotor rightTop = null;
+    public DcMotor rightBottom = null;
+    public CRServo duckWheelRight = null;
+    public CRServo duckWheelLeft = null;
+    public DcMotor arm = null;
+    public Servo claw = null;
 
-@Autonomous(name="Blue20SecAuto", group="Autonomous")
-//@Disabled
-public class Blue20SecAuto extends LinearOpMode {
-    /* Declare OpMode members. */
-    private ElapsedTime runtime = new ElapsedTime();
-    private DcMotor leftTop;
-    private DcMotor leftBottom;
-    private DcMotor rightTop;
-    private DcMotor rightBottom;
-    private DcMotor arm;
-    private Servo claw;
+    public static final double CLAW_OPEN    =  0 ;
+    public static final double CLAW_CLOSED  =  1 ;
+    public static final double CLAW_MID     =  0.5 ;
 
-    public void setup() {
+    //    public static final double ARM_UP_POWER    =  0.45 ;
+    //    public static final double ARM_DOWN_POWER  = -0.45 ;
 
+    /* local OpMode members. */
+    HardwareMap hardwareMap  =  null;
+    Telemetry telemetry      = null;
+
+    public ElapsedTime runtime  = new ElapsedTime();
+
+    /* Constructor */
+    public RobotConfigTrish() {
+
+    }
+
+    /* Initialize standard Hardware interfaces */
+    public void init(HardwareMap ahwMap, Telemetry tltry) {
+        // Save reference to Hardware map
+        hardwareMap = ahwMap;
+        // Set up telemetry
+        telemetry = tltry;
+
+                // Define and Initialize Motors
         leftTop = hardwareMap.get(DcMotor.class, "leftTop");
         leftBottom = hardwareMap.get(DcMotor.class, "leftBottom");
         rightTop = hardwareMap.get(DcMotor.class, "rightTop");
         rightBottom = hardwareMap.get(DcMotor.class, "rightBottom");
         arm = hardwareMap.get(DcMotor.class, "arm");
         claw = hardwareMap.get(Servo.class, "claw");
+        duckWheelRight = hardwareMap.get(CRServo.class, "duckWheel1");
+        duckWheelLeft = hardwareMap.get(CRServo.class, "duckWheel2");
 
         leftTop.setDirection(DcMotor.Direction.REVERSE);
         leftBottom.setDirection(DcMotor.Direction.REVERSE);
         rightTop.setDirection(DcMotor.Direction.FORWARD);
         rightBottom.setDirection(DcMotor.Direction.FORWARD);
 
+        // Set all motors to zero power
+        drive(0, 0);
+
+        // Reset all encoders.
         arm.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         leftTop.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         leftBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightTop.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
 
-        // Send telemetry message to signify robot waiting;
-        telemetry.addData("Status", "Initiated");    //
-
-        telemetry.update();
+        // Initialize ALL installed servos.
+        claw.setPosition(CLAW_CLOSED);
     }
 
-    @Override
-    public void runOpMode() {
-
-        setup();
-
-        // Wait for the game to start (driver presses PLAY)
-        waitForStart();
-
-        claw.setPosition(1.0);
-        sleep(20000);
-
-        // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
-
-        raiseArm(150);
-
-        strafeToPosition(890);
-
-        raiseArm(-145);
-
-        // Step 1:  Drive forward for 3 seconds
-        /*arm.setPower(0.3);
-        runtime.reset();
-        while (opModeIsActive() && (runtime.seconds() < 1.0)) {
-            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-            telemetry.update();
-        }
-
-        strafeRight(0.3);
-
-        runtime.reset();
-        while (opModeIsActive() && (runtime.seconds() < 4.0)) {
-            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-            telemetry.update();
-        }
-
-        // Step 2:  Stop
-        strafeRight(0); //Stops motors
-        arm.setPower(0);
-
-
-        telemetry.addData("Path", "Complete");
-        telemetry.update();
-        sleep(1000);*/
-    }
-
-
-    private void driveToPosition(int drivePosition) {
+    public void driveToPosition(int drivePosition) {
         int distance = leftBottom.getCurrentPosition() + drivePosition;
+        telemetry.addData("distance", drivePosition);
+        telemetry.addData("leftBottom Current position", leftBottom.getCurrentPosition());
+        telemetry.addData("leftBottom New position", distance);
         leftBottom.setTargetPosition(distance);
         leftTop.setTargetPosition(distance);
         rightBottom.setTargetPosition(distance);
         rightTop.setTargetPosition(distance);
+        telemetry.update();
+
+//        sleep(2000);
+
 
         leftBottom.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         leftTop.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         rightBottom.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         rightTop.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
+        telemetry.update();
         drive(0.5, 0.5);
 
         while (leftBottom.isBusy()) {
+            telemetry.addData("leftBottom Actual position", leftBottom.getCurrentPosition());
+            telemetry.update();
+
             // do nothing
 //            telemetry.addData("Arm position", arm.getCurrentPosition());    //
 //            telemetry.update();
 
         }
+//        sleep(2000);
+
         // 6. Turn off the motor
         drive(0, 0);
     }
 
-    private void strafeToPosition(int strafePosition) {
+    public void strafeToPosition(int strafePosition) {
         leftTop.setTargetPosition(leftTop.getCurrentPosition() - strafePosition);
         leftBottom.setTargetPosition(leftBottom.getCurrentPosition() + strafePosition);
         rightTop.setTargetPosition(rightTop.getCurrentPosition() + strafePosition);
@@ -185,18 +177,18 @@ public class Blue20SecAuto extends LinearOpMode {
         drive(0, 0);
     }
 
-    private void left(double power) {
+    public void left(double power) {
         try {
-            leftTop.setPower(power);
             leftBottom.setPower(power);
+            leftTop.setPower(power);
         } catch (Exception ex) {
         }
     }
 
-    private void right(double power) {
+    public void right(double power) {
         try {
-            rightTop.setPower(power);
             rightBottom.setPower(power);
+            rightTop.setPower(power);
         } catch (Exception ex) {
         }
     }
@@ -207,17 +199,47 @@ public class Blue20SecAuto extends LinearOpMode {
     }
 
 
-    private void raiseArm(int position) {
+    public void strafeLeft(double speed) {
+        telemetry.addData("Strafe speed", speed);
+
+
+        leftTop.setPower(-speed * 0.75);
+        leftBottom.setPower(speed * 0.75);
+        rightTop.setPower(speed * 0.75);
+        rightBottom.setPower(-speed * 0.75);
+    }
+
+    public void strafeRight(double speed) {
+        telemetry.addData("moveRight speed", speed);
+
+
+        strafeLeft(-speed);
+    }
+
+//    public void controlArm(double speed) {
+//        arm.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+//        arm.setPower(speed);
+//    }
+
+    public void raiseArmToPosition(int position) {
+        if (position == 0) {
+            return; // do nothing
+        }
+
 //        arm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         //       arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         telemetry.addData("Distance", position);    //
         telemetry.addData("Starting Arm position", arm.getCurrentPosition());    //
+
         int armPosition = arm.getCurrentPosition() + position;
         telemetry.addData("New Arm position", armPosition);    //
 
-        arm.setTargetPosition(armPosition);
+        if (armPosition < 0) {
+            return; // do nothing
+        }
 
+        arm.setTargetPosition(armPosition);
 
         arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 //        telemetry.addData("Arm position", arm.getCurrentPosition());    //
@@ -226,11 +248,22 @@ public class Blue20SecAuto extends LinearOpMode {
         while (arm.isBusy()) {
             // do nothing
         }
-        telemetry.addData("Actual Arm position", arm.getCurrentPosition());    //
-//        arm.setPower(0);
-        sleep(2000);
-        telemetry.addData("Fallen Arm position", arm.getCurrentPosition());    //
-        telemetry.update();
-        sleep(1000);
     }
-}
+
+    public void ducks(double duckSpeed) {
+        duckWheelRight.setPower(duckSpeed);
+    }
+
+    public void setClawOpen() {
+      claw.setPosition(CLAW_OPEN);
+    }
+
+    public void setClawClosed() {
+        claw.setPosition(CLAW_CLOSED);
+    }
+
+    public void setClawMid() {
+        claw.setPosition(CLAW_MID);
+    }
+ }
+
