+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/autonomous/BlueFreight.java
@@ -27,11 +27,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.ftc_2021_2022;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022.autonomous;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -40,26 +40,26 @@ import com.qualcomm.robotcore.util.ElapsedTime;
  * This file illustrates the concept of driving a path based on time.
  * It uses the common Pushbot hardware class to define the drive on the robot.
  * The code is structured as a LinearOpMode
- *
+ * <p>
  * The code assumes that you do NOT have encoders on the wheels,
- *   otherwise you would use: PushbotAutoDriveByEncoder;
- *
- *   The desired path in this example is:
- *   - Drive forward for 3 seconds
- *   - Spin right for 1.3 seconds
- *   - Drive Backwards for 1 Second
- *   - Stop and close the claw.
- *
- *  The code is written in a simple form with no optimizations.
- *  However, there are several ways that this type of sequence could be streamlined,
- *
+ * otherwise you would use: PushbotAutoDriveByEncoder;
+ * <p>
+ * The desired path in this example is:
+ * - Drive forward for 3 seconds
+ * - Spin right for 1.3 seconds
+ * - Drive Backwards for 1 Second
+ * - Stop and close the claw.
+ * <p>
+ * The code is written in a simple form with no optimizations.
+ * However, there are several ways that this type of sequence could be streamlined,
+ * <p>
  * Use Android Studios to Copy this Class, and Paste it into your team's code folder with a new name.
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@Autonomous(name="Red Warehouse", group="Autonomous")
+@Autonomous(name = "Blue Freight", group = "Autonomous")
 //@Disabled
-public class RedWarehouse extends LinearOpMode {
+public class BlueFreight extends LinearOpMode {
 
     /* Declare OpMode members. */
     private ElapsedTime runtime = new ElapsedTime();
@@ -67,8 +67,11 @@ public class RedWarehouse extends LinearOpMode {
     private DcMotor leftBottom;
     private DcMotor rightTop;
     private DcMotor rightBottom;
+    //private DcMotor duckWheel;
     private DcMotor arm;
     private Servo claw;
+    //private CRServo duckWheel;
+
 
     public void setup() {
 
@@ -76,8 +79,10 @@ public class RedWarehouse extends LinearOpMode {
         leftBottom = hardwareMap.get(DcMotor.class, "leftBottom");
         rightTop = hardwareMap.get(DcMotor.class, "rightTop");
         rightBottom = hardwareMap.get(DcMotor.class, "rightBottom");
+        //duckWheel = hardwareMap.get(DcMotor.class, "duckWheel");
         arm = hardwareMap.get(DcMotor.class, "arm");
         claw = hardwareMap.get(Servo.class, "claw");
+        //duckWheel = hardwareMap.get(CRServo.class, "duckWheel");
 
         leftTop.setDirection(DcMotor.Direction.REVERSE);
         leftBottom.setDirection(DcMotor.Direction.REVERSE);
@@ -89,6 +94,8 @@ public class RedWarehouse extends LinearOpMode {
         leftBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightTop.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+
+//        arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         // Send telemetry message to signify robot waiting;
         telemetry.addData("Status", "Initiated");    //
@@ -104,43 +111,37 @@ public class RedWarehouse extends LinearOpMode {
         // Wait for the game to start (driver presses PLAY)
         waitForStart();
 
+        // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
+
+        // Step 1:  Drive forward for 3 seconds
+
         claw.setPosition(1.0);
         sleep(1000);
 
-        // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
+        //ducks(1);
 
         raiseArm(150);
 
-        strafeToPosition(890);
+//        leftTop.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+//        leftBottom.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+//        rightTop.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+//        rightBottom.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        raiseArm(-145);
+        int drivePosition = 860;
 
-        // Step 1:  Drive forward for 3 seconds
-        /*arm.setPower(0.3);
-        runtime.reset();
-        while (opModeIsActive() && (runtime.seconds() < 1.0)) {
-            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-            telemetry.update();
-        }
+        driveToPosition(drivePosition);
 
-        strafeRight(0.3);
+        sleep(500);
 
-        runtime.reset();
-        while (opModeIsActive() && (runtime.seconds() < 4.0)) {
-            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-            telemetry.update();
-        }
+        claw.setPosition(0);
+        sleep(500);
 
-        // Step 2:  Stop
-        strafeRight(0); //Stops motors
-        arm.setPower(0);
+        raiseArm(50);
+        driveToPosition(-125);
 
+        strafeToPosition(1650);
 
-        telemetry.addData("Path", "Complete");
-        telemetry.update();
-        sleep(1000);*/
     }
-
 
     private void driveToPosition(int drivePosition) {
         int distance = leftBottom.getCurrentPosition() + drivePosition;
@@ -162,6 +163,7 @@ public class RedWarehouse extends LinearOpMode {
 //            telemetry.update();
 
         }
+
         // 6. Turn off the motor
         drive(0, 0);
     }
@@ -211,7 +213,7 @@ public class RedWarehouse extends LinearOpMode {
 
     private void raiseArm(int position) {
 //        arm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        //       arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+ //       arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         telemetry.addData("Distance", position);    //
         telemetry.addData("Starting Arm position", arm.getCurrentPosition());    //
@@ -236,5 +238,61 @@ public class RedWarehouse extends LinearOpMode {
         sleep(1000);
     }
 
+    /*private void ducks(int duckSpeed) {
+        duckWheel.setPower(duckSpeed);
+    }*/
 
+    /*leftTop.setPower(0.4);
+    leftBottom.setPower(0.4);
+    rightTop.setPower(0.4);
+    rightBottom.setPower(0.4);
+    while (opModeIsActive() && (runtime.seconds() < 1.25)) {
+        telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
+        telemetry.update();
+    }
+
+    strafeRight(0.3);
+
+    runtime.reset();
+    while (opModeIsActive() && (runtime.seconds() < 1.0)) {
+        telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
+        telemetry.update();
+    }
+
+    // Step 2:  Stop
+    strafeRight(0); //Stops motors
+
+
+    telemetry.addData("Path", "Complete");
+    telemetry.update();
+    sleep(1000);
+}
+*/
+    public void moveLeft(double speed) {
+        telemetry.addData("moveLeft speed", speed);
+
+        double rb_speed_var = 1;
+        if (speed < 0) {
+            rb_speed_var = -1.0 / speed;
+        }
+        double rb_speed = speed * rb_speed_var;
+        telemetry.addData("Strafe speed", speed);
+        telemetry.addData("rb_speed_var", rb_speed_var);
+        telemetry.addData("rb_speed", rb_speed);
+
+        leftTop.setPower(-speed);
+        leftBottom.setPower(speed);
+        rightTop.setPower(speed);
+        rightBottom.setPower(-rb_speed);
+
+        telemetry.addData("LT Speed", leftTop.getPower());
+        telemetry.addData("LB Speed", leftBottom.getPower());
+        telemetry.addData("RT Speed", rightTop.getPower());
+        telemetry.addData("RB Speed", rightBottom.getPower());
+    }
+
+    public void moveRight(double speed) {
+        telemetry.addData("moveRight speed", speed);
+        moveLeft(-speed);
+    }
 }
