+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc_2021_2022/autonomous/BlueWarehouseForward.java
@@ -27,11 +27,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.ftc_2021_2022;
+package org.firstinspires.ftc.teamcode.ftc_2021_2022.autonomous;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -40,26 +39,26 @@ import com.qualcomm.robotcore.util.ElapsedTime;
  * This file illustrates the concept of driving a path based on time.
  * It uses the common Pushbot hardware class to define the drive on the robot.
  * The code is structured as a LinearOpMode
- * <p>
+ *
  * The code assumes that you do NOT have encoders on the wheels,
- * otherwise you would use: PushbotAutoDriveByEncoder;
- * <p>
- * The desired path in this example is:
- * - Drive forward for 3 seconds
- * - Spin right for 1.3 seconds
- * - Drive Backwards for 1 Second
- * - Stop and close the claw.
- * <p>
- * The code is written in a simple form with no optimizations.
- * However, there are several ways that this type of sequence could be streamlined,
- * <p>
+ *   otherwise you would use: PushbotAutoDriveByEncoder;
+ *
+ *   The desired path in this example is:
+ *   - Drive forward for 3 seconds
+ *   - Spin right for 1.3 seconds
+ *   - Drive Backwards for 1 Second
+ *   - Stop and close the claw.
+ *
+ *  The code is written in a simple form with no optimizations.
+ *  However, there are several ways that this type of sequence could be streamlined,
+ *
  * Use Android Studios to Copy this Class, and Paste it into your team's code folder with a new name.
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@Autonomous(name = "Red Freight Duck Wheel", group = "Autonomous")
+@Autonomous(name="BlueWarehouseFoward", group="Autonomous")
 //@Disabled
-public class RedFreightDucks extends LinearOpMode {
+public class BlueWarehouseForward extends LinearOpMode {
 
     /* Declare OpMode members. */
     private ElapsedTime runtime = new ElapsedTime();
@@ -67,10 +66,8 @@ public class RedFreightDucks extends LinearOpMode {
     private DcMotor leftBottom;
     private DcMotor rightTop;
     private DcMotor rightBottom;
-    private CRServo duckWheel1;
     private DcMotor arm;
     private Servo claw;
-
 
     public void setup() {
 
@@ -80,7 +77,6 @@ public class RedFreightDucks extends LinearOpMode {
         rightBottom = hardwareMap.get(DcMotor.class, "rightBottom");
         arm = hardwareMap.get(DcMotor.class, "arm");
         claw = hardwareMap.get(Servo.class, "claw");
-        duckWheel1 = hardwareMap.get(CRServo.class, "duckWheel1");
 
         leftTop.setDirection(DcMotor.Direction.REVERSE);
         leftBottom.setDirection(DcMotor.Direction.REVERSE);
@@ -92,8 +88,6 @@ public class RedFreightDucks extends LinearOpMode {
         leftBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightTop.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightBottom.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-
-//        arm.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         // Send telemetry message to signify robot waiting;
         telemetry.addData("Status", "Initiated");    //
@@ -109,79 +103,67 @@ public class RedFreightDucks extends LinearOpMode {
         // Wait for the game to start (driver presses PLAY)
         waitForStart();
 
-        // Make sure good grip on block
         claw.setPosition(1.0);
         sleep(1000);
 
-        //ducks(1);
+        // Step through each leg of the path, ensuring that the Auto mode has not been stopped along the way
 
-        // Raise arm to top level
         raiseArm(150);
 
-        // Drive to alliance hub
-        int drivePosition = 860;
+        strafeToPosition(890);
+
+        int drivePosition = 500;
         driveToPosition(drivePosition);
-        sleep(500);
 
-        // Release cube
-        claw.setPosition(0);
-        sleep(500);
+        raiseArm(-145);
 
-        // Raise arm to not get stuck
-        raiseArm(50);
+        // Step 1:  Drive forward for 3 seconds
+        /*arm.setPower(0.3);
+        runtime.reset();
+        while (opModeIsActive() && (runtime.seconds() < 1.0)) {
+            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
+            telemetry.update();
+        }
 
-        // Backup before driving away
-        driveToPosition(-125);
+        strafeRight(0.3);
 
-        // Drive to wall by storage
-        strafeToPosition(1650);
-        sleep(500);
+        runtime.reset();
+        while (opModeIsActive() && (runtime.seconds() < 4.0)) {
+            telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
+            telemetry.update();
+        }
+
+        // Step 2:  Stop
+        strafeRight(0); //Stops motors
+        arm.setPower(0);
 
 
-        driveToPosition(-300);
-//        sleep(5000);
-
-        double duckSpeed = 1;
-        ducks(duckSpeed);
-//        sleep(5000);
-
-        driveToPosition(300);
-
+        telemetry.addData("Path", "Complete");
+        telemetry.update();
+        sleep(1000);*/
     }
+
 
     private void driveToPosition(int drivePosition) {
         int distance = leftBottom.getCurrentPosition() + drivePosition;
-        telemetry.addData("distance", drivePosition);
-        telemetry.addData("leftBottom Current position", leftBottom.getCurrentPosition());
-        telemetry.addData("leftBottom New position", distance);
         leftBottom.setTargetPosition(distance);
         leftTop.setTargetPosition(distance);
         rightBottom.setTargetPosition(distance);
         rightTop.setTargetPosition(distance);
-        telemetry.update();
-
-        sleep(2000);
-
 
         leftBottom.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         leftTop.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         rightBottom.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         rightTop.setMode(DcMotor.RunMode.RUN_TO_POSITION);
 
-        telemetry.update();
         drive(0.5, 0.5);
 
         while (leftBottom.isBusy()) {
-            telemetry.addData("leftBottom Actual position", leftBottom.getCurrentPosition());
-            telemetry.update();
-
             // do nothing
 //            telemetry.addData("Arm position", arm.getCurrentPosition());    //
 //            telemetry.update();
 
         }
-        sleep(2000);
-
         // 6. Turn off the motor
         drive(0, 0);
     }
@@ -209,16 +191,16 @@ public class RedFreightDucks extends LinearOpMode {
 
     private void left(double power) {
         try {
-            leftBottom.setPower(power);
             leftTop.setPower(power);
+            leftBottom.setPower(power);
         } catch (Exception ex) {
         }
     }
 
     private void right(double power) {
         try {
-            rightBottom.setPower(power);
             rightTop.setPower(power);
+            rightBottom.setPower(power);
         } catch (Exception ex) {
         }
     }
@@ -255,62 +237,4 @@ public class RedFreightDucks extends LinearOpMode {
         telemetry.update();
         sleep(1000);
     }
-
-    private void ducks(double duckSpeed) {
-        duckWheel1.setPower(duckSpeed);
-    }
-
-    /*leftTop.setPower(0.4);
-    leftBottom.setPower(0.4);
-    rightTop.setPower(0.4);
-    rightBottom.setPower(0.4);
-    while (opModeIsActive() && (runtime.seconds() < 1.25)) {
-        telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-        telemetry.update();
-    }
-
-    strafeRight(0.3);
-
-    runtime.reset();
-    while (opModeIsActive() && (runtime.seconds() < 1.0)) {
-        telemetry.addData("Path", "Leg 1: %2.5f S Elapsed", runtime.seconds());
-        telemetry.update();
-    }
-
-    // Step 2:  Stop
-    strafeRight(0); //Stops motors
-*/
-
-    /*telemetry.addData("Path", "Complete");
-    telemetry.update();
-    sleep(1000);*/
-
-/*
-    public void moveLeft(double speed) {
-        telemetry.addData("moveLeft speed", speed);
-
-        double rb_speed_var = 1;
-        if (speed < 0) {
-            rb_speed_var = -1.0 / speed;
-        }
-        double rb_speed = speed * rb_speed_var;
-        telemetry.addData("Strafe speed", speed);
-        telemetry.addData("rb_speed_var", rb_speed_var);
-        telemetry.addData("rb_speed", rb_speed);
-
-        leftTop.setPower(-speed);
-        leftBottom.setPower(speed);
-        rightTop.setPower(speed);
-        rightBottom.setPower(-rb_speed);
-
-        telemetry.addData("LT Speed", leftTop.getPower());
-        telemetry.addData("LB Speed", leftBottom.getPower());
-        telemetry.addData("RT Speed", rightTop.getPower());
-        telemetry.addData("RB Speed", rightBottom.getPower());
-    }
-
-    public void moveRight(double speed) {
-        telemetry.addData("moveRight speed", speed);
-        moveLeft(-speed);
-    }*/
 }
